airflow/requirements.txt
regex==2023.12.25
pyarrow==15.0.0
psycopg2-binary==2.9.9
numpy==1.23.5
pandas==2.0.3
apache-airflow-providers-google==10.14.0
